message: new class counterBox for reuse
